Match bombplant_update hitType range compare width
(u32) cast on the (hitType - 0xe) range check emits cmplwi instead of
signed cmpwi. bombplant_update 99.81 -> 99.97% (remaining: stack slot
permutation).

+++ src/main/dll/dll_01A9_bombplant.c
@@ -403,7 +403,7 @@ void bombplant_update(void* obj)
             {
                 Obj_StartModelFadeIn(obj, 0x12c);
             }
-            else if (hitType - 0xe <= 1 || hitType == 0x11)
+            else if ((u32)(hitType - 0xe) <= 1 || hitType == 0x11)
             {
                 Sfx_PlayFromObject(obj, SFXmv_curtainloop16);
                 hitX = hitX + playerMapOffsetX;
